Move the data type parsing logic to Metadata\Property class
