+++ src/Tuicha/Metadata/Property.php
@@ -2,8 +2,10 @@
 
 namespace Tuicha\Metadata;
 
+use Notoj\Annotation\Annotation;
 use Notoj\Annotation\Annotations;
 use Notoj\ReflectionProperty;
+use RuntimeException;
 
 class Property
 {
@@ -19,21 +21,22 @@ class Property
     protected $required = false;
     protected $isPublic = true;
 
-    public function __construct($phpName, $mongoName = null, $reflection = null)
+    public function __construct($phpName, $mongoName = null, ReflectionProperty $reflection = null)
     {
+        $this->phpName   = $phpName;
+        $this->mongoName = $mongoName ?: $phpName;
+        $this->type      = $this->type ?: new DataType($mongoName === '_id' ? 'id' : '');
+
         if ($reflection) {
             $annotations = $reflection->getAnnotations();
-            $phpName     = $reflection->getName();
-            $mongoName   = $annotations->has('field') ? $annotations->getOne('field')->getArg(0) : $phpName;
+            $this->phpName   = $reflection->getName();
+            $this->mongoName = $annotations->has('field') ? $annotations->getOne('field')->getArg(0) : $this->phpName;
             if ($annotations->has('id')) {
-                $mongoName = '_id';
+                $this->mongoName = '_id';
             }
 
             $this->parseReflection($reflection, $annotations);
         }
-        $this->phpName   = $phpName;
-        $this->mongoName = $mongoName ?: $phpName;
-        $this->type      = new DataType($mongoName === '_id' ? 'id' : '');
     }
 
     public function parseReflection(ReflectionProperty $reflection, Annotations $annotations)
@@ -42,6 +45,7 @@ class Property
         $this->isPublic    = $reflection->isPublic();
         $this->required    = $annotations->has('required');
         $this->validations = $this->getAnnotationArguments($annotations->get('validate'));
+        $this->type        = $this->parseDataType($annotations);
     }
 
     /**
@@ -92,6 +96,52 @@ class Property
     public function isPublic()
     {
         return $this->isPublic;
+    }
+
+    /**
+     * Get the type definition from a single Annotation.
+     *
+     * @return array
+     */
+    protected function getDataTypeFromAnnotation(Annotation $annotation)
+    {
+        $type = new DataType($annotation->getName());
+
+        switch ($annotation->getName()) {
+        case 'class':
+            $type->addData('class', strtolower($annotation->getArg()));
+            break;
+        case 'array':
+            try {
+                $type->addData('element', $this->getDataTypeFromAnnotation($annotation->getArg()));
+            } catch (RuntimeException $e) {
+            }
+            break;
+        case 'reference':
+            foreach ($annotation->getArgs() as $name => $value) {
+                $type->addData($name, $value);
+            }
+            break;
+        }
+
+        return $type;
+    }
+
+    /**
+     * Returns the data type definition for a property
+     *
+     * @param Annotations $annotations  Property's annotations object
+     *
+     * @return array
+     */
+    protected function parseDataType(Annotations $annotations)
+    {
+        $types = 'int,integer,float,double,array,bool,boolean,string,object,class,type,id,reference';
+        if ($annotation = $annotations->getOne($types)) {
+            return $this->getDataTypeFromAnnotation($annotation);
+        }
+
+        return $this->mongoName === '_id' ?  new DataType('id') : $this->type;
     }
 
     public function getType()
